RPM: createrepo project property pages
Added project property pages. Shifted "Deltas" options from the plugin
preferences to the project specific preferences. Also added a createrepo
project nature so that eclipse knows when to display the project
property pages (for createrepo natures & .repo files).

Changed when the "content" folder is to be created only when importing
RPMs or executing the createrepo command, so that when a .repo file is
created in a different project, a content folder won't appear.

Added an abstract class for createrepo property pages.

Change-Id: I9c728ea6dcd0b0020be7d3ff676d448a3746b85c
Signed-off-by: Neil Guzman <nguzman@redhat.com>
Reviewed-on: https://git.eclipse.org/r/19323
Reviewed-by: Alexander Kurtakov <akurtako@redhat.com>
IP-Clean: Alexander Kurtakov <akurtako@redhat.com>
Tested-by: Alexander Kurtakov <akurtako@redhat.com>