ctf: Add unittests for the CTF Lexer and Parser.
This is a first draft of unittests for the CTF Lexer and Parser.
The coverage of these tests is still incomplete. More unittests
are to come.

Some tests are commented and need to fix the grammars.
The plan is to fix them in future commits to limits the
amount of refactoring.

Change-Id: I2d8e06560314ef22710985915b726a74974bbb1c
Signed-off-by: Etienne Bergeron <etienne.bergeron@gmail.com>
Reviewed-on: https://git.eclipse.org/r/19485
Tested-by: Hudson CI
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
IP-Clean: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>