TMF: Added a method in UI tests to delay main thread until trace is opened
In o.e.l.tmf.ui.tests, the main thread often need to be put to sleep while a
trace element is being opened. The sleep time was put to 500 ms because that's
what showed best results. But sometimes, unit tests fail because it takes a bit
more than that. Instead of putting longer delays, we loop a certain number of
times until the trace is opened or after a certain timeout.

Change-Id: I662d951115eacbb3fa5b17cfe534a8e5c2a85eb4
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/20018
Tested-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
IP-Clean: Matthew Khouzam <matthew.khouzam@ericsson.com>