lttng: Check the state system query results more rigorously
Defensive programming dictates we should check for integrity,
correctness, etc. at different layers. And not just once and
assume it's safe at every other levels.

In this case, if we had a trace with a sched_process_free event
arriving first for a new TID, the CFV would assume the Exec_name
attribute would be already set, which was not the case.

Change-Id: I0489937be07282c236978a07e48c4221160bb1d9
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/21490
Tested-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
IP-Clean: Patrick Tasse <patrick.tasse@gmail.com>