tmf: Store the ID in the state system itself
Many components already refer to state systems with an ID, might
as well store it in the state system object itself for better
encapsulation.

Beyond the SSID pun, I like having it *not* be getId(), because that
quickly becomes confusing with IAnalysisModule.getId() (even though
those two values are usually the same).

Change-Id: I51babc177f54b4f8af66544e6c75be53da56b6c4
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/21540
Tested-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>