Systemtap: Fix errors with opening .stp files.
Prevent editor errors caused by opening out-of-workspace .stp
files, or by using OpenFileAction to open any .stp file:

-PathEditorInput's getAdapter(Class) now returns a non-null adapter,
allowing out-of-workspace files to open properly with it.
(Resolves EBZ #428149.)

-OpenFileAction now behaves like Eclipse's OpenLocalFileAction,
meaning that it uses default behaviour to choose the appropriate
input & editor type. This also means that File->Open File... and
the "Open" button on the Systemtap toolbar now choose the same
kind of editor to use for a .stp file.

Change-Id: If526eb4459efce5caa906e6cf18298d579751942
Signed-off-by: Andrew Ferrazzutti <aferrazz@redhat.com>
Reviewed-on: https://git.eclipse.org/r/22039
Tested-by: Hudson CI