TMF: Add possibility to cancel update thread in XY line viewer
When the update takes a long time to run (or when it is updated too frequently)
many threads may be running at the same time and the results are not accurate.
The update is now done in its own thread class that can be cancelled and the
updateData method has a monitor to cancel when necessary.

Change-Id: I1430e0d2329b896d85c48326dcd2bef4a013d7a9
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/22460
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
IP-Clean: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>