gdbtrace: Prevent hang when a lot of commands are sent to GDB
This can happen when commands are sent to GDB on the UI thread
and the GDB traces console gets full. The UI thread is responsible of
doing the "flush" of the console buffer so if the UI thread waits for
the flush it will wait for itself forever. The solution is to not send
GDB commands on the UI thread.

Bug: 415728
Change-Id: I6eacb9680f4966183a24c5ad915e4ddff256f40d
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/23019
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
IP-Clean: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>