ctf: Specify a timeout for the CTF testsuite download
The default timeout of the Ant "get" task is reasonable, but the
default for "exec" (which we use for pulling the CTF testsuite) feels
like 5 minutes. This makes building offline, or behind a proxy, a bit
more painful.

Shorten the delay to 20 seconds.

Change-Id: Ic0751330c4c747fee4344d549257f7a6ab556d0d
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/23070
Tested-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
IP-Clean: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>