TMF: Add rebuild() method to abstract time graph view
It is useful when some action in the view requires the view to rebuild the
list of entries, for a same trace.

This patch also adds a getPresentationProvider() method to the time graph
view so concrete classes may make changes to the presentation provider.

Change-Id: Ida104642607b98d4d285bcc54a8e86dd73602506
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/23147
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
IP-Clean: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Hudson CI