tmf: Fix time graph combo freezing when expanding parent of many items
When expanding an item in the time graph combo, the expanded state of
every child item needs to be synched between the tree and the time
graph, due to inconsistent behavior between platforms. When the parent
has hundreds of child items, this can take several seconds due to the
inefficient searching for the item corresponding to a time graph entry
in both the tree and the time graph, for getting and setting the
expanded state.

For the tree, instead of gettting the expanded state of each item from
the tree viewer using the model element, the set of expanded elements
is retrieved once and the expanded state can be deduced by the presence
or absence of a model element in this set. Furthermore the synching of
expanded state is only performed for child items that have children
themselves.

For the time graph, the item model is changed to use a map from time
graph entry to time graph item, which greatly reduced the time necessary
to find the item corresponding to an entry. The expanded item index is
also stored directly in the item which avoids the need to iterate to
find an expanded time graph entry's index.

Change-Id: Icefc393be8a5e2ee1a98b98a2c6b126da7490035
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/23817
Tested-by: Hudson CI