tmf: Fix batch import doesn't respect user-chosen trace name on copy
When the user changes the trace display name in the batch import wizard
(for example to resolve a name conflict), this name is not used when
importing if the copy option is selected and the trace is a folder.

This prevents the trace type to be set correctly on a single trace
import, and prevents name-conflicted traces to be imported properly.

Change-Id: I3dd04ca5deeaf5d4f9327ef13e7c1fdd5dcd2b62
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/24122
Tested-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
IP-Clean: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
IP-Clean: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Tested-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>