TMF: Add trace stub for TMF unit tests
This new trace type uses a custom XML definition and converts custom XML
events to TmfEvents, where the fields and their values are defined through the
XML fields. To illustrate, here is the content of each type of event:

CustomXmlEvent: fields = "a | b", values = "1 | abd"

would transpose to this

Development trace: a = 1, b = "abd"

This type of trace will be useful to unit test analyses involving LTTng kernel
and UST traces for instance.

Change-Id: I41ac7206c1a54088c998d551a4174f2184b05f5b
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/21265
Tested-by: Hudson CI