tmf: Bug 424964: Drag and dropping a linked trace doesn't create a link
Unless the DROP_LINK operation is specifically requested by the user
(e.g. by holding Alt key), the source resource type (linked or not) is
now maintained during the drag and drop operation.

The handling is also changed to allow drag and drop of a resource in a
Tracing project that is not part of the tracing model (e.g. a file or
folder that is outside of the Traces and Experiments sub-trees).

A possible NullPointerException is fixed when the link creation fails.

Change-Id: I3b80ee508bfecb4023a1b52b2befdda7003d3794
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/24475
Tested-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>