Bug 350291: Gcov annotations should appear in the C/C++ editor.
Gcov has a separate editor for displaying code annotations. To better
integrate with C/C++ developer tooling, the annotations should appear in
the same editor being used to view the source code.

The Gcov annotation model renders for an open ICEditor if the
corresponding project has been launched using the Gcov profiling launch
delegate. The model also renders when a user attempts to open a Gcov
data file (.gcno, .gcda). Once a build-related action (clean, build) is
performed on a project, the annotation model is disabled for all files
within that project.

Change-Id: Ia7966207261cfbb376902b9099cfc78962f553ca
Reviewed-on: https://git.eclipse.org/r/24876
Reviewed-by: Roland Grunberg <rgrunber@redhat.com>
Tested-by: Roland Grunberg <rgrunber@redhat.com>