ctf: Make CTFTrace and its reader classes AutoCloseable
Basically just replace their dipose() method with close(). This
small change allows us to benefit from the compiler warnings for
when a resource is allocated locally in a method but never disposed.

Which then allows to find leaks. Many leaks ! Many tests were leaking
CTFTraceReaders and the like, but the biggest offender is probably the
validation methods in LttngKernelTrace and LttngUstTrace: trying to
validate a non-CTF trace could potentially leak a number of file
descriptors equal to the number of files in that trace(!)

Change-Id: I7e54416b608992a14a7bfd3cd29ffb8dc3cee898
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/24986
Tested-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>