ctf: Introducing the ByteArrayDefinition
This patch introduces a special acceleration that will make sequences
and arrays read faster when it's an array of bytes or a  string, as the
CTF spec specifies that they are aligned on 8 bits and always 8 bits wide
per character.

Sequences and arrays can either be ArrayDefinitions or
ByteArrayDefintions.

This patch gives a 20% acceleration in sequential reading.

Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Change-Id: I7933268b7f9271f74f7d59339ea9b8343fdb021e
Reviewed-on: https://git.eclipse.org/r/26084
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Tested-by: Hudson CI