tmf: Add TmfEventTableColumn class
Introduce the concept of "event table column" objects. This
object will contain:
- The initial column configuration (previously in "ColumnData")
- The getItemString() "functor" defining how to populate the cell
  in this column for a given event.
- The "field ID" to allow searching/filtering to keep working as
  before. Eventually, the filtering mechanism could be reworked
  to just use getItemString() directly.

The base framework now defines 5 columns, that are available for
all trace types (whose getter methods are currently in ITmfEvent):
- Timestamp
- Source*
- Event Type
- Reference*
- Contents (fields)

(* = could eventually be removed, subject to future discussion)

Every trace type can now specify which columns it wants to display,
which may or may not include the default-provided ones.

This is the first step towards decoupling the trace type, event table,
and column order and number. This will eventually allow displaying all
the possible trace type columns in an experiment of heterogeneous traces.

Change-Id: I4ffdb0564114a92a58a271a7a99d3ca43fa57e31
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/30468
Tested-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>