lttng-control: Workaround for stderr problem in with SSH shell
The RSE ssh shell service uses JCraft's ChannelShell class which
doesn't provide the error stream. This causes problems when parsing
the error output for a special behaviour. For example, when a
the kernel provider is not available the command lttng list -k returns
an error message in the error stream "Kernel provider not available"
and has a non-zero return value. However, this error is ingored in the
Control view, since some users only use UST.

For MI integration, it's important that stdout and stderr is provided
separately.  This patch works for the MI integration, unless there
is mixture of stdout and stderr output for the case that a subset
of the command is successful and some fail (e.g. enable multiple
events). In that case the output has to treated in a special way.

Change-Id: I22f6aa98f36c80602d0f73c0d9a25ae5131ef16b
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/31570
Tested-by: Hudson CI
Reviewed-by: Jonathan Rajotte Julien <jonathan.rajotte-julien@ericsson.com>
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>