statesystem: Remove duplicate empty state creation
The transient state already creates a "null" value for every state
attribute that is added. No need for the attribute tree to do an
explicit insertion immediately after (which gets ignored any way).

That attribute tree should also keeps its paws off the transient state,
but that's a different problem.

Change-Id: Ib77e16d67375e65c495b9ff0b73c5cee78eddf02
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/33076
Tested-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>