tmf: Do not use Display.sleep in tests
Similar to bug 415992 and change 25443. The display needs to be woken up
because it's possible to get in a state where nothing will wake up the UI
thread (events, messages), especially if the tests are being run unattended.
Instead of using the highly implementation-dependent solution of doing a
Display.asyncExec (which only works on Mac, not Windows), we can simply not put
the display to sleep and just have the thread sleep. This is safer because
Display.sleep by definition should not wake up until an event is received which
we cannot guarantee will happen.

Change-Id: Ib2282bcd71c9376d7bbdb084c094892146410ff7
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/33257
Tested-by: Hudson CI
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Tested-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>