Bug 469717: Work around socket read timeout in docker-client.
attachContainer in docker-client uses the regular client which by
default has a fixed socket read timeout. When attaching to a container,
an arbitrary amount of time may pass between reads so when a timeout
occurs, simply re-attach in the reading thread.

Note that this only happens for TCP connections. The UNIX socket
implementation seems to have data traveling consistently while the TCP
implementation (Apache HttpClient) likely takes care of this
functionality itself, leaving the socket unused for longer.

Change-Id: I0ecbf64ba145bbea7cebfd9094272ffe9d371cee
Reviewed-on: https://git.eclipse.org/r/53795
Tested-by: Hudson CI
Reviewed-by: Roland Grunberg <rgrunber@redhat.com>
Tested-by: Roland Grunberg <rgrunber@redhat.com>