Bug 494521 - NPE when New Docker Connection dialog is open on Windows
Prevent returning connection settings when the DOCKER_HOST
environment variable was not found.

On environments where Docker Toolbox was installed, there
might not be any DOCKER_HOST, etc. env variables set, but there
is a 'DOCKER_TOOLBOX_INSTALL_PATH' env var, which means that
the exit code of the script is '0' and from there, the
'createDockerConnectionSettings' is called.
This method noew returns 'null' if the DOCKER_HOST property
is missing.

Change-Id: Ia19b31782fcd714b9c6549b01801129f5295936f
Signed-off-by: Xavier Coulon <xcoulon@redhat.com>
Reviewed-on: https://git.eclipse.org/r/73588
Tested-by: Hudson CI
Reviewed-by: Jeff Johnston <jjohnstn@redhat.com>