Refresh images view after removing temporary image.
When pushing some image 'baz' to registry 'foo.bar:5000', we tag an
image temporarily as 'foo.bar:5000/baz' to facilitate the push. At the
very end this tag is removed, so the images view should be refreshed so
the image doesn't incorrectly show up until a manual refresh.

Related: Bug 471661
Change-Id: Ied3474a9810af0641d53b6b23911d5d08916733e
Reviewed-on: https://git.eclipse.org/r/73751
Tested-by: Hudson CI
Reviewed-by: Jeff Johnston <jjohnstn@redhat.com>