Restore equals() and hashCode() methods for DockerConnection.
104b675 removed equals() and hashCode() because they originally took the
connection name into account. As a result, changing the connection name
would generate new refresh threads since the map that stored these used
the connection as the key.

The equals() and hashCode() methods should be determined by the host
path / certificates used and not on any given name. Delegate to the
TCP/UNIX implementation of IDockerConnectionSetting and remove calls to
their parent (BaseConnectionSettings) which checks the name. Also ensure
that duplicate connections aren't added.

Change-Id: I9800ff12b51cd3f44e95ebefdef1579b0bdbdffc
Reviewed-on: https://git.eclipse.org/r/77550
Tested-by: Hudson CI
Reviewed-by: Roland Grunberg <rgrunber@redhat.com>