Bug 500638 - Properties do not show information about selected image/container in respective tab
Fixed the value returned by 'getContributorId()' in
DockerImagesView and DockerContainersView to re-enable linking selection
with properties view (was a regression introduced in fix for bug 499750
(Add commands in Image Hierarchy view).

Added tests to verify that the properties view displays the expected tab
upon selection in DockerExplorerView, DockerContainersView, DockerImagesView
and DockerImageHierarchyView.

Change-Id: I61038304c86b2dda02c46728d69e5dd351d8dcf9
Signed-off-by: Xavier Coulon <xcoulon@redhat.com>
Reviewed-on: https://git.eclipse.org/r/80221
Tested-by: Hudson CI
Reviewed-by: Jeff Johnston <jjohnstn@redhat.com>