Bug 508657 - NoStackTrace in RunDockerImageLaunchConfigurationDelegate.launch
Externalized the deserialization of port bindings, converting a `List<String>`
containing serialized values into a `Map<String, List<IDockerPortBinding>>`.
Added tests on the `LaunchConfigurationUtilsTest` class.

Also:
- fixed a bug where the list of images displayed in the combo
would not be populated.
- fixed a bug where the current, selected connection in the Docker Explorer view
would be used instead of the selected connection in the launch editor, causing
some errors if the currently selected connection was not opened.

Change-Id: I034da6dd21298cb24ea57c881bd05bc5f2b539c9
Signed-off-by: Xavier Coulon <xcoulon@redhat.com>
Reviewed-on: https://git.eclipse.org/r/86378
Tested-by: Hudson CI
Reviewed-by: Jeff Johnston <jjohnstn@redhat.com>