Add support for running on Windows.
On Windows, java.io.File converts Unix paths to Windows ones so we need
to create our own class that subclasses java.io.File, but preserves the
logic of getAbsolutePath(), getName(), etc.

On Windows, we won't have direct access to the container's IP address
since the Docker daemon will exist within a VM. As a result we need to
ensure the container's debug port is forwarded to the VM and then we
need to search for any IP address bound to an active network interface
that is listening on the port we've specified. To further reduce the
likelihood of false positives, the port should be randomized.

Change-Id: If56b995da930f43100954af2e3e83abb99bb2d8d
Reviewed-on: https://git.eclipse.org/r/98939
Reviewed-by: Roland Grunberg <rgrunber@redhat.com>
Tested-by: Roland Grunberg <rgrunber@redhat.com>